app/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
click==8.1.7
dnspython==2.6.1
elastic-transport==8.12.0
elasticsearch==8.6.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.104.1
frozenlist==1.4.1
h11==0.14.0
idna==3.6
multidict==6.0.5
orjson==3.6.8
pydantic==1.10.2
redis==4.5.1
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.9.0
urllib3==2.2.1
uvicorn==0.24.0
uvloop==0.17.0
yarl==1.9.4
python-dotenv==1.0.1
Django==4.0.4
psycopg2-binary==2.9.9
uWSGI==2.0.23
django-split-settings==1.2.0
django-debug-toolbar==4.2.0
django-cors-headers==4.3.1
requests~=2.31.0
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-exporter-jaeger==1.21.0
opentracing==2.4.0